fix: 取得可觀看類別修正，引入service
